// estbm-frontend/src/app/models/stage.model.ts
export interface SoutenanceEtudiantSlotDto {
entreprise: any;
  etudiantId: number;
  date: string;       // ISO date
  heureDebut: string; // "HH:mm:ss"
  heureFin: string;   // "HH:mm:ss"
  sujet: string;
}